# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API documentation for the Translation Management Service built with Laravel",
        "contact": {
            "email": "shayan@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Shayan"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "shayan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and get token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "shayan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List all translations (with optional filters)",
                "operationId": "03e0745fb3bed90fe04f46a1395fa4c1",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag (e.g. mobile, web)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "mobile"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by translation key",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "greeting"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Filter by translated content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Welcome"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "greeting"
                                            },
                                            "translations": {
                                                "properties": {
                                                    "en": {
                                                        "type": "string",
                                                        "example": "Hello"
                                                    },
                                                    "fr": {
                                                        "type": "string",
                                                        "example": "Bonjour"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "mobile",
                                                    "web"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "711dfe2c7f0dbd38f67d0f3e7e07f5b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "translations"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "welcome"
                                    },
                                    "translations": {
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "example": "Welcome"
                                            },
                                            "fr": {
                                                "type": "string",
                                                "example": "Bienvenue"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "web",
                                            "mobile"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/{id}": {
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update an existing translation",
                "operationId": "a5cb050be20f56c1ce385df0166adef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "greeting"
                                    },
                                    "translations": {
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "example": "Hello"
                                            },
                                            "fr": {
                                                "type": "string",
                                                "example": "Bonjour"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "mobile",
                                            "web"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated"
                    },
                    "404": {
                        "description": "Translation not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export all translations in JSON format",
                "operationId": "9f227c61c766f0f5cacdb54bab6df4a2",
                "responses": {
                    "200": {
                        "description": "Translations exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "en": {
                                        "welcome": "Welcome",
                                        "goodbye": "Goodbye"
                                    },
                                    "fr": {
                                        "welcome": "Bienvenue",
                                        "goodbye": "Au revoir"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Translations",
            "description": "Translations"
        }
    ]
}